IMPALA-1599: Use BE-specific table descriptors for plan fragments
When sending a plan fragment to all backends, the table descriptor
object would reference a T*Table, which is the frontend's Thrift
representation of a table. However, that data structure is much more
verbose than it needs to be for backend execution. Size and
serialization cost are very important for heavily partitioned tables in
large clusters.

This patch introduces T*TableDescriptor data structures for all of our
physical table types, as well a THdfsPartitionDescriptor for HDFS table
partitions. These structs have exactly the information in them required
to execute a query against them, and no more.

The effect of this patch and its predecessor is to reduce the size of a
plan fragment that included the descriptor for a fully-scanned 1k
column, 7.3k partition table by 37%.

Change-Id: Id710d054f8ecf1c8a6970894f89e4a6e771822a3