IMPALA-1599: Allow data stream senders to start before receivers
This patch allows data stream senders to start before their receivers,
by forcing the sender to block during its first transmission until the
receiver is ready.

This is done by establishing a 'rendezvous' between sender and receiver,
based on a Promise whose return type is the receiver itself. All senders
that can not find their receiver wait on that Promise. If the receiver
arrives promptly, it will trigger the Promise and the senders can all
continue. If the receiver does not arrive within 60s, the sender will
time out.

TODO: When a receiver isn't present, the sender will fail quietly (this
was always the case, see comments). We should look at fixing this,
although the false-positives need working around.

This commit also adds a 'tombstone' cache for recently cancelled
receivers.

When a data-stream sender can't find its receiver (during
TransmitData()), it waits for up to 60s by default for the receiver to
arrive. However, this can lead to unnecessary waiting in the case when
the receiver has already been closed, and will never show up.

In most cases, this should not happen, but there are two that are worthy
of some attention:

  * Cancellation happens while TransmitData() is in-flight, and the
    receiver is closed, but the sender does not know until the RPC times
    out (we don't have cancellable RPCs which would be the natural way
    to implement this).
  * The receiver is closed forcibly through some other path than the
    usual EOS-receipt from all its senders. This happens, for example,
    when an agg. node with a limit is above an exchange - the agg. node
    will close the exchange once it considers all the rows it needs to
    be read. Then we can have the situation where the sender is trying
    to TransmitData() to close its connection to the exchange, but the
    exchange has already closed.

The result is that any code that waits for a backend to be completed
will take longer due to hitting the timeout path.

To address this, this patch adds a cache of recently expired streams to
the data stream manager. Any TransmitData() call that is targeting a
stream in that cache will fail fast. Entries are kept in the cache for
30s, and are evicted collaboratively by other closing senders during
CloseSender().

Change-Id: If9a90f6980ab2cbd4b59b54f737da73c6af88f1a