IMPALA-1599: Allow data stream senders to start before receivers
This patch allows data stream senders to start before their receivers,
by forcing the sender to block during its first transmission until the
receiver is ready.

This is done by establishing a 'rendezvous' between sender and receiver,
based on a Promise whose return type is the receiver itself. All senders
that can not find their receiver wait on that Promise. If the receiver
arrives promptly, it will trigger the Promise and the senders can all
continue. If the receiver does not arrive within 10s, the sender will
time out.

Change-Id: If9a90f6980ab2cbd4b59b54f737da73c6af88f1a