IMPALA-1599: Make Prepare() async for plan fragments
When codegen is enabled, Prepare() for plan fragments can be relatively
expensive. However, we call it on the critical path for query setup,
during the ExecRemoteFragment() RPC issued by the coordinator. This can
lead to high startup latencies for queries, particularly those with many
fragments.

This patch moves Prepare() to the fragment exec thread. This allows the
coordinator to start many fragments in quick succession. The main change
this involves is pushing responsibility for reporting an error during
Prepare() to a one-time invocation of ReportStatusCallback() so that the
coordinator will find out about the error.

Change-Id: Ie39737dc419d7708dd881e68d1035e05d3256d19