IMPALA-1599: Issue fragment exec RPCs from a thread pool
This patch changes the dispatch of ExecRemoteFragment() calls to use a
thread pool rather than a parallel executor. The advantages are that the
threads are only started once per query (rather than once per level of
the plan tree), and the coordinator does not wait for one set of plan
fragments to be started before trying to start the next if there are
threads free.

This is the payoff after the changes to allow for asynchronous fragment
start-up in DataStreamSender and FragmentExecState in previous patches.

Some of the expensive initialisation of BackendExecStates (from large
parameter copies) is moved into a new Init() call so that it can be done
in parallel as well. This patch also cleans up BackendExecState from the
struct-like class that it was to a properly written class where members
are hidden behind getters etc. since BackendExecState was growing to
include more and more of its own logic.

A new thread pool is added to the ExecEnv for dispatching
ExecRemoteFragment() RPC calls. This uses a new CallableThreadPool class
that makes implementing general-purpose thread pools a bit easier.

Change-Id: Ifa9ba7212093f83c5960b9509ca9d2a31d8fde56