IMPALA-1599: Issue fragment exec RPCs from a thread pool
This patch changes the dispatch of ExecRemoteFragment() calls to use a
thread pool rather than a parallel executor. The advantages are that the
threads are only started once, and the coordinator does not wait for one
set of plan fragments to be started before trying to start the next if
there are threads free.

Some of the expensive initialisation of BackendExecStates is moved into
a new Init() call so that it can be done in parallel as well.

Change-Id: Ifa9ba7212093f83c5960b9509ca9d2a31d8fde56