Optional Impala Toolchain
This patch allows to optionally enable the new Impala binary
toolchain. For now there are now major version differences in the
toolchain dependencies and what is currently kept in thirdparty.

To enable the toolchain, export the variable IMPALA_TOOLCHAIN to the
folder where the binaries are available. Make sure to source
impala-config before building.

Change-Id: If2dac920520e4a18be2a9a75b3184a5bd97a065b