IMPALA-1599: Add 'tombstone' cache for recently cancelled receivers
When a data-stream sender can't find its receiver (during
TransmitData()), it waits for up to 60s by default for the receiver to
arrive. However, this can lead to unnecessary waiting in the case when
the receiver has already been closed, and will never show up.

In most cases, this should not happen, but there are two that are worthy
of some attention:

  * Cancellation happens while TransmitData() is in-flight, and the
    receiver is closed, but the sender does not know until the RPC times
    out (we don't have cancellable RPCs which would be the natural way
    to implement this).
  * The receiver is closed forcibly through some other path than the
    usual EOS-receipt from all its senders. This happens, for example,
    when an agg. node with a limit is above an exchange - the agg. node
    will close the exchange once it considers all the rows it needs to
    be read. Then we can have the situation where the sender is trying
    to TransmitData() to close its connection to the exchange, but the
    exchange has already closed.

The result is that any code that waits for a backend to be completed
will take longer due to hitting the timeout path.

To address this, this patch adds a cache of recently expired streams to
the data stream manager. Any TransmitData() call that is targeting a
stream in that cache will fail fast. Entries are kept in the cache for
30s, and are evicted asynchronously by a dedicated thread.

Change-Id: Ibb2ad0019be7fcacde54e218a6c4b888ca8651d1