IMPALA-2642: Fix potential deadlock when there are too many subscribers
Statestore::DoSubscriberUpdate(), as its last action, schedules the next
update for that subscriber by calling OfferUpdate(). However, it does so
under subscriber_lock_, which it shouldn't, because OfferUpdate() can
take the same lock if the subscribers queue is too large, in order to
erase the current subscriber (i.e. to load shed).

This patch just moves OfferUpdate() out of the lock scope. Writing a
test is hard because this is a hard situation to reproduce - you need
10k subscribers that are misbehaving, then one that works well, and
together those typically blow past the ulimit for files on a machine (as
well as being stressful).

Change-Id: I66df2a27843cd65f9d60fcda4a3589c1beea07f1