CDH-33292: Add support for dynamic partition pruning (FE)
This commit adds support for dynamic partition pruning in the frontend.
During planning, the frontend computes a set of filters that are
constructed by join operators and are applied at scan operators in order
to reduce the partitions than need to be scanned. The filters are identified
from equi-join predicates by traversing the single-node plan tree in a
top-down fashion.

A query option, termed enable_dynamic_partition_pruning, is added to
enable/disable DPP (disabled by default). When DPP is enabled, the
output of EXPLAIN is modified to include information about the runtime
filters that are constructed/applied during DPP. Also, an event is added
to the query timeline to track the time spent in the planner while
computing DPP.

Testing:
Functional planner tests are added.

Change-Id: Id79a38313051d95da32c897b176a40d26b0dda1d