IMPALA-3023: Check return from AllocateScratchBloomFilter()
RuntimeFilterBank::AllocateScratchBloomFilter() may return NULL, but
this was not checked in the short-circuit path for UpdateFilterFromLocal().

This patch also ensures there is no race to call SetBloomFilter()
between short-circuit publication in UpdateFilterFromLocal(), and
PublishGlobalFilter(). This ensures that arrival_time_ is only set once,
which is used for reporting in the runtime profile.

Change-Id: I22bf059d518855a507ab47285e80659c64e92e2f